requirements.txt
python>=3.8

PyPDF2>=3.0.0           
python-docx>=0.8.11    
openai>=0.27.0          

pydantic>=1.10.0        

transformers>=4.25.0    
torch>=1.13.0           
scikit-learn>=1.0.0    

python-dotenv>=0.21.0   

pytest>=7.0.0          
pytest-cov>=3.0.0       

requests>=2.28.0       
numpy>=1.23.0           